// src/common/types/Profile.ts
export type Profile = {
  id: string
  avatar_url?: string
  first_name?: string
  middle_name?: string
  last_name?: string
  email_address: string
  phone_number?: string
  street_address?: string
  city?: string
  state?: string
  zip?: number
  company?: string
  occupation?: string
  website?: string
  birthday?: string
}